feat: add SEO-optimized meta tags, structured JSON-LD data & SEO score
