+++ package.json
@@ -16,6 +16,7 @@
         "@testing-library/user-event": "^12.8.3",
         "@xmtp/browser-sdk": "^4.1.0",
         "axios": "^1.8.4",
+        "cheerio": "^1.1.2",
         "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
         "connectkit": "^1.9.1",
